Match objective elementwise metric structure
