+++ src/metric/elementwise_metric.h
@@ -1,12 +1,11 @@
 /**
- * Copyright 2015-2026, XGBoost Contributors
+ * Copyright 2026, XGBoost Contributors
  * \file elementwise_metric.h
- * \brief Shared declarations and CPU kernels for elementwise metrics.
+ * \brief Typed elementwise metric kernels and CPU implementations.
  */
 #ifndef XGBOOST_METRIC_ELEMENTWISE_METRIC_H_
 #define XGBOOST_METRIC_ELEMENTWISE_METRIC_H_
 
-#include <cmath>    // for log1p, sqrt
 #include <cstddef>  // for size_t
 #include <numeric>  // for accumulate
 #include <vector>   // for vector
@@ -14,50 +13,23 @@
 #include "../common/kernel.h"            // for KernelRegistration
 #include "../common/optional_weight.h"   // for OptionalWeights
 #include "../common/threading_utils.h"   // for ParallelFor1d
-#include "metric_common.h"               // for CheckRowWeights, PackedReduceResult
-#include "xgboost/base.h"                // for bst_float
+#include "metric_common.h"               // for PackedReduceResult
 #include "xgboost/context.h"             // for Context, DeviceOrd
 #include "xgboost/data.h"                // for MetaInfo
 #include "xgboost/host_device_vector.h"  // for HostDeviceVector
 #include "xgboost/linalg.h"              // for UnravelIndex
 
-namespace xgboost::metric {
-struct EvalRowRMSE {
-  char const* Name() const { return "rmse"; }
-
-  XGBOOST_DEVICE bst_float EvalRow(bst_float label, bst_float pred) const {
-    bst_float diff = label - pred;
-    return diff * diff;
-  }
-  static double GetFinal(double esum, double wsum) {
-    return wsum == 0 ? std::sqrt(esum) : std::sqrt(esum / wsum);
-  }
-};
-
-struct EvalRowRMSLE {
-  char const* Name() const { return "rmsle"; }
-
-  XGBOOST_DEVICE bst_float EvalRow(bst_float label, bst_float pred) const {
-    bst_float diff = std::log1p(label) - std::log1p(pred);
-    return diff * diff;
-  }
-  static double GetFinal(double esum, double wsum) {
-    return wsum == 0 ? std::sqrt(esum) : std::sqrt(esum / wsum);
-  }
-};
-
-namespace elementwise {
-template <typename Policy>
+namespace xgboost::metric::elementwise {
+template <typename EvalFn>
 struct EvalKernel {
   using Signature = PackedReduceResult(Context const*, HostDeviceVector<float> const&,
-                                       MetaInfo const&, Policy);
+                                       MetaInfo const&, EvalFn);
 };
 
 namespace detail {
-template <typename Policy>
+template <typename EvalFn>
 PackedReduceResult EvalCpu(Context const* ctx, HostDeviceVector<float> const& preds,
-                           MetaInfo const& info, Policy policy) {
-  CheckRowWeights(info);
+                           MetaInfo const& info, EvalFn eval) {
   auto labels = info.labels.HostView();
   auto predts = preds.ConstHostSpan();
   common::OptionalWeights weights{info.weights_.ConstHostSpan()};
@@ -72,7 +44,7 @@ PackedReduceResult EvalCpu(Context const* ctx, HostDeviceVector<float> const& pr
     for (std::size_t i = block.begin(), n = block.end(); i < n; ++i) {
       auto [sample_id, target_id] = linalg::UnravelIndex(i, labels.Shape());
       float weight = weights[sample_id];
-      float residue = policy.EvalRow(labels(sample_id, target_id), predts[i]) * weight;
+      float residue = eval(labels(sample_id, target_id), predts[i]) * weight;
       sum_score += residue;
       sum_weight += weight;
     }
@@ -88,12 +60,11 @@ PackedReduceResult EvalCpu(Context const* ctx, HostDeviceVector<float> const& pr
 }
 }  // namespace detail
 
-template <typename Policy>
+template <typename EvalFn>
 auto RegisterEvalCpu() {
-  using Kernel = EvalKernel<Policy>;
-  return common::KernelRegistration<Kernel>{DeviceOrd::kCPU, &detail::EvalCpu<Policy>};
+  using Kernel = EvalKernel<EvalFn>;
+  return common::KernelRegistration<Kernel>{DeviceOrd::kCPU, &detail::EvalCpu<EvalFn>};
 }
-}  // namespace elementwise
-}  // namespace xgboost::metric
+}  // namespace xgboost::metric::elementwise
 
 #endif  // XGBOOST_METRIC_ELEMENTWISE_METRIC_H_
